requirements.txt
langchain_together 
python-dotenv
linkedin-api
tiktoken